Rename interface file
