+++ src/background/details-view-controller.ts
