+++ src/electron/adapters/null-details-view-controller.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { DetailsViewController } from 'background/details-view-controller-type';
+import { DetailsViewController } from 'background/details-view-controller';
 
 export class NullDetailsViewController implements DetailsViewController {
     public setupDetailsViewTabRemovedHandler(handler: (tabId: number) => void): void {}
